Support for burst tokens in Vrouter
Currently the numbher of hold flow entries is limited by a configurable
parameter "HoldCount" in Vrouter. If for some reason, agent does not get
scheduled, there is a chance that the hold queue gets filled and Vrouter
starts dropping the packets. To accomodate these type of flow drops,
burst tokens are introduced in Vrouter. On top of the hold count, once
the hold queue is full, Vrouter releases fixed number of tokens for
every fixed amount of time. This accomodates burstly behaviour of the
traffic and also Agent not getting scheduled by OS. In a steady state,
only packets equivalent to hold count would be in HoldQueue. Once the
hold queue is full, as there are fixed number of tokens are getting
released for every given period of time, these tokens take care of the
burst.

The "vrouter" command is modified to configure three more parameters.
"Total number of allowed burst tokens", time interval with which burst
tokens need to be released and total number of burst token per every
interval. The free burst tokens are displayed in "flow -r" command and
configure tokens are displayed in "vrouter --info" command.

Change-Id: I2754109f9847d3ec1ab601ab0aa601ec09c155ff
closes-bug: #1631862