Address potential memory leak and double-free in vhost_rx_handler.
If a shared sk_buff was passed into vhost_rx_handler() it would clone
it, and return RX_HANDLER_ANOTHER. However, it failed to tell the
caller about the new sk_buff. The end result was that the new sk_buff
would leak, and the old one would have its ref count decremented
twice.

The fix is to update the sk_buff ** passed in by the caller.

Change-Id: Ifb29f245831cc7c9855c727442e5e7bd469123ed
Closes-Bug: #1679929