Calcaulte udp source port for multicast pkt using right hash
While calculating the UDP source port for a multicast packet, the hash
needs to be on Ethernet header and VRF Id. For packet arriving from VM
and Tor, there are no extra header other than transport headers. So the
hash is calculated rightly for these packets. But for packets arriving
from other compute node, there are additional header other than
transport headers. The hash is wrongly calculated in this case on thsese
additional headers leading to different hash for the same Ethernet
headers, resulting in different UDP sourcd port.

The offset is correctly calculated and right ethernet header is passed
to hash function as fix.

Change-Id: I93ae0d96e536875988da6a525cdba4783c1e4d6e
closes-bug: #1635046