Infrastructure to support aging for L2 entries
To support aging, there should be

. a way to indicate the activity of an entry
. a way to easily read/monitor the activity

For the first requirement, add packet statistics to each entry. If
learning is enabled, the statistics will increment for both source
and the destination L2 entries. Else, the statistics will increment
only for the destination L2 entry, since no lookup is done for the
source.

For the second requirement, bridge table is now exposed to applications
as a memory block that can be mapped and read like any other memory.
This way, the application does not have to send sandesh request for
each and every bridge entry just to see whether there is any activity
in that entry.

Change-Id: I12bb766853580805519bc04d19790bc62853fbef
Partial-Bug: #1636413