Mirroring for packets coming from agent
For packets that are transmitted from agent to a specific interface,
we are setting the packet type to AGENT. Unless the packet type is
IP, we do not do a COPY-SIP operation for mirroring (which basically
selects the correct IP to be used for the mirror packet), with the
side effect that the packet will have wrong source IP (typically the
vhost IP) and hence will be dropped in the destination vRouter that
hosts the mirror VM.

As a fix, do not set the packet type to AGENT for such packets. Such
packets, when mirrored will have the SIP as the destination IP of the
original packet, similar to the case of packets from fabric, since
port mirroring for such packets will happen from the destination VM
that is hosted in the same compute node as the one where agent is
running. Also, when we select the correct source IP for mirroring,
we will make sure that the SIP and DIP will not be the same.

Change-Id: I0882c83702fa9d7ee1cc2ceea3d5a553bb401ef7
Closes-Bug: #1637225