Avoid flow processing for neighbor solicitation/advertisements
A unicast neighbor solicitation/response undergoes a flow lookup.
In many cases the source IP address that is used will be the link
local IP which will get dropped because of RPF checks. Consistent
with the way we treat ARP requests, we will now make sure that
the solicitation and advertisements do not undergo flow lookup.

Change-Id: Ib2ab306b0bdd6f201c35faa1918f97324238be1e
Closes-Bug: #1643906