Hold packets while doing ECMP resolution
In a typical ECMP use case, by the time packet destined to the ECMP
destination hits the flow, the ECMP index to be used to select the
best among equals would already have been decided by the agent and
programmed accordingly in the vRouter. However, in some not so common
scenarios, the first packet can be trapped to agent for resolving the
nexthop to be used among the nexthops that form the ECMP destination.

In such cases, we should hold and trap the packet rather than just
trapping so that first packet loss can be avoided.

Change-Id: I09e785254ee0ef187099152e2808b8a67850d31d
Closes-Bug: #1648027