DPDK: Vrouter crash in pkt_drop_stats
In some cases, vr_packet* may not be populated in the mbuf.
If such a packet is freed, it may cause a crash when its accounted
since tries to access vif* structure.

Fix is to add vif argument to vr_dpdk_pfree() to ensure that a valid
vif is always present and the packet is accounted for that vif

Change-Id: If0a46e4e7261e2021a40a0bab086834cde2bf7a2
Closes-Bug:1646358