Invalidate the ingress interface of pkt while flushing from hold q
When the Agent writes the flow action, if the hold q exists, the packets
from hold q are reinjected to regular data path. Before doing this, the
packet's ingress interfaces pointer is validated against VIF db
pointer, as a sanity check, to ensure that VIF is same before
holding the packet and while flushing. If not same, the packet
is dropped, as VIF is a not the same any more. If VIF is deleted, while
the packet is held in hold Q, vp_if member of pkt holds an invalid
pointer leading to wrong VIF memory access. In this case,
vif_pcpu_drop_stats is wrongly accessed leading to crash.

As a fix vp_if of pkt is invalidated before invoking pkt_drop_stats() so
that we dont increment per VIF drop stats but only global drop stats.

Change-Id: I59cd72c2634a3629a92e2b154736a750c9ab32c9
closes-bug: #1649749