Broadcast vrouter operations on netlink
This patch introduces the netlink broadcasting of operations on
several vrouter resources. For instance, when a route is added by a
client, the vrouter broadcasts the request on netlink.

We currently broadcast 'add' and 'delete' operations on the resources:
route, qos, interface, mirror, vxlan, vrf_assign, nexthop, mpls.

This allows the userspace to be notified by any modifcation on these
resources, in the same manner it is notified by the Linux network
stack thanks to 'ip monitor' for instance.

Change-Id: Ifed994e552b35faa771d4b6dad6da7eea1cecdfb
Depends-On: Ieb868404fc0d614d5fa7cb26db4d931e802d49a0
Closes-bug: #1650316