HW Queueing for DPDK based vRouter
After the QOS module in vRouter makes the decision to send a
packet to a particular hardware queue, the platform specific
parts need to interface with the individual drivers and make
sure that packets go to the desired queue in the NIC. The
vRouter module for the Linux Kernel already does this.

In the case of DPDK, each lcore is bound to a specific queue.
Hence, if a packet comes in one lcore and needs to go to a
queue that is bound to another lcore, the packet needs to
go to that particular lcore. To achieve that, we create rings
for every queue other than the queue that is bound to the
lcore in each forwarding lcores. Also, the number of
'vr_dpdk_queue', which is the vRouter representation of the
queue, for each PHYSICAL interface that supports more than one
queue needs to be equal to the number of queues configured by
user/ agent. There also needs to be a quick way to reach the
dpdk queue corresponding to the queue marked in the packet.

Change-Id: If6dd5afaf7bf55f35b5fbc3b38a311e28442a964
Partial-Bug: #1617938