Always proxy with "MYMAC" if such a flag is set in the interface
In certain use cases, it is desired to do layer 3 processing for
all packets that are egressing a virtual interface. In such cases,
regardless of the destination and their mac stitching status, mac
requests for the destination should always be proxied with "MYMAC"
so that packets are subjected to layer 3 processing.

Define a new control flag at the vif layer that agent can use to
request the above mentioned behavior.

Partial-Bug: #1651030

Conflicts:
	include/vr_interface.h
	utils/vif.c

Change-Id: I805b4ed1e4d86c19c8e579e6c515dcff9b53a6a5