DPDK: GRO and GSO support
Vrouter performs Tx and Rx TCP segmentation offload only in the kernel mode currently.
It largely leverages kernel APIs for achieving this. However, in case of DPDK, the library
lacks segmentation support due to which DPDK based vrouter could not do the offloads.

This commit overcomes the above limitation and adds receive offload and send offload
support to DPDK. As a pre-requisite to this, it also adds mbuf chaining support and virtio
mergeable buffer support.

Closes-Bug: 1592935

Change-Id: I124b1cf0ff45465e1180727bf1666b468dceaf19