Do not blindly bridge link local neighbor solicitations
The existing code that handles neighbor solicitations, bridges
all requests to link local destinations. This behavior does not
work for gateway link local address since

. a broadcast request to the link local address will not make
  to the agent

. a unicast request to the link local address will not be handled
  by the agent.

Agent does not handle neighbor solicitations since vRouter is
expected to respond back with the proper mac address based on
the flags in the route.

Hence, make all link local neighbor solicitations to go through
vRouter's logic of proxying and flooding based on flags set on
the route.

Change-Id: I4f23d1c980e887d044776d49bbc71b9faaf2f0a9
Closes-Bug: #1657483

Set the 'Router' bit if proxying with Router MAC

For ICMPv6 advertisements that carry the Router MAC, set the
'Router' bit in the advertisements.

Change-Id: Iccc5d1ca025c71f55bfc85c6e12f4cf6944a9dfc
Closes-Bug: #1657483