Remove the constraint of V6 prefix length to be multiple of 8
While adding the V4 and V6 routes to Mtrie, bits that are set in prefix
beyond the mask length are reset to 0. While doing this for V6 prefixes,
it is assumed that V6 mask len is byte boundary. If not, thre prefix
corresponding to those bits are also reset to zero.

As a fix, this constraint has been removed

Change-Id: I2c61ea9829d37b78d8570411bc3db02c0ef1cded
closes-bug: #1657766