DPDK: Vrouter crash in pkt_drop_stats
In some cases, vr_packet* may not be populated in the mbuf.
If such a packet is freed, it may cause a crash when its
accounted since tries to access vif* structure.

Fix is to add vif argument to vr_dpdk_pfree() to ensure that
a valid vif is always present and the packet is accounted for that vif

Change-Id: I792aa331a835043396ddcc32bfd965fb5df05af1
Closes-Bug:1646358