Double tag the packet on Vlan interface if required
Currently packet is tagged on Vlan interfce in the tx routine with the
outgoing vlan tag. If the packet is already tagged, it is not double
tagged any more. This is because, if the packet is processed in L3 Encap
NH, the rewrite data would have already contained the required tag, and
tagging again in tx routine makes it a double tag. But if the packet is
received from VM with multiple tags, packet very likey gets processed in
L2 ecnap NH and the outer tag needs to be tagged in tx routine. This
tagging needs to happen even if the packet is already tagged with inner
tag, as outgoing packet is intended to be multi tagged.

When the packet is received on VMI with multiple tags, the upper tag is
used to identify the subinterface. This vlan tag is consumed and the
inner vlan tag, if any, is stored in fmd while parsing the packet. In
case of VRF assignment, the vlan tag is used to assign the VRF and is
stored in fmd. So this fmd's vlan identifies whether packet is received
on VMI as double tagged or not. This is used to for the tag on tx
routine of the interface.

Change-Id: I556d48e4e534fe163b9707c19809825673db6c04
closes-bug: #1662129