Assign RPF call back function for L2 encap nexthop
Currently the RPF validation call back, nh_validate_src, is not assigned
to L2 Encap Nexthop. In case of PBB, this is resulting in not
identifying the MAC MOVE. In PBB code, the mac move is identified by
calling this RPF callback function, if exists, which returns whether the
source is valid source or not.

As a fix, the existing L3 encap NH RPF call back is assigned to L2 encap
as well. Also this call back is supposed to mark FMD whether the packet
is from Root or not using NH_FLAG_ETREE_ROOT. This functionality is also
added in this call back.

Change-Id: Ic63d57c211a40ae1f1167b89c73dce9f559384cc
closes-bug: #1661637