Allow updating Bridge entry flags
When the bridge entry is modified from Agent, the entry is modified with
the requests Flags as well. To ensure atomic modification of the flags,
new flags are copied as long as entry's flags are not modifed by any one
else. To achieve this, __sync_bool_compare_and_swap() is used. But this
is invoked not with old flags as is. This is resulting in not updating
the new flags to beidge entry.

In case of PBB, if control word is toggled in VN, it results in new
flags to be written to Bridge entry. Not updating the bridge entry flags
is resulting in control word not taking effect.

As a fix, the atomic update is invoked with right old values.

closes-bug: #1662141

Change-Id: I2101d48c18ac402e71fc0099c0d968857d3f70b3