For PBB Multicast BMac validation dont use pkt's Multicast flag
When PBB Broadcast/Multicast traffic is replicated in Multicast tree,
the destination Bmac is derived from ISID. In the receiving Vrouters,
this BMac is validated using the same derivation, in L2 Encap NH, if the
packet is marked as VP_FLAG_MUTLICAST. Packet is marked as
VP_FLAG_MUTLICAST if the destination MAC of ethernet header is brodcast
or multicacast. In unknown unicast case the packet would not be marked
with multicast flag as it is unicast mac address, but the destination
bmac would be multicast pbb mac as it is replicated using multicast
tree. This results in wrong validation of Pbb mac, as PBB dest mac
multicast mac and packet is not marked with multicast flag.

To address this, if PBB destination MAC, which is stored in fmd, is
multicast/broadcast bmac, it is compared against bmac using isid
derivation, without using pkt flag.

closes-bug: #1661959

Change-Id: If1bd18b7181de0f640f593d5e9f097b2fb25e96f