Move flow lookup of fragments to common hash table infrastructure
With the older infrastructure, for every absent entry there will
be a linear lookup done in the complete overflow table, which is
very expensive. Such an expensive lookup can result in push back
of packet processing to queues further up in the path and can
ultimately result in packet drops if the processing rate is not
able to match up with the arrival rate.

The newer hash table infrastructure avoids linear lookup in the
overflow table, instead simply iterating on the list maintained
/bucket to identify present/absent entries.

This fix should ease the situation explained in the bug report,
where packets were dropped in the queues maintained by the OS
because vRouter was too busy looking up overflow table.

Change-Id: Ie11b947acdc5adec263d7b7d98de0d59d3d86f44
Closes-Bug: #1659350