Convert port to network order while making udp header in mirroring
The nh_udp_tunnel_helper() expects the port argument to be in network
byte order and it does not convert them to network order by itself. The
mirroring udp tunnel nexthop invokes this helper routine with ports as
host order. This leads to source port going out of range and dest port
not being identified as intended port in receiving machines.

As a fix, the helper routine is invoked by converting the ports to
network byte order

Change-Id: I32575cf7de4bde072216f22683f6b6926f6933be
closes-bug: #1654633