Null checks for bridge table locks
While doing Mac learning in data path, there is a lock used to ensure
that same Source MAC is not learnt in multiple cores. This lock is
maintained per VIF and memory is allocated and freed while MAC learning
is toggled on VIF. If MAC learning is disabled on VIF, lock data is made
NULL in VIF immediately and freeing of this memory is done using
RCU call back. There is a possibility of bridge learning code accessing
bridge lock data while this happens.  This bridge lock data can become
NULL any time while bridge learning in progress. To ensure that we dont
operate on NULL bridge lock data it is taken into local variable at the
start and rest of the operations are done on this local variable.

Change-Id: Ia3ee89c79ddaa42e047e457831ab5bd8c6873d8d
closes-bug: #1675264