Dont create Fragment table for every soft reset
Rightnow fragment table, which is a hash table, is created every time
soft reset is done by Agent, without  validating whether hashtable
already exists or not. This is resulting in memory leak as old hash
table is never deleted. As a fix, fragment table is created only if
it is non NULL

Change-Id: If54d997967da33cefc15b99efb5f7619bdf88132
closes-bug: #1675309