Audit memory frees against allocs
In order to detect potential double frees and possible memory
corruption, write a signature, the object type and the state
of the memory at the beginning of the memory block during
allocation. At the time of free, audit the memory block against
these values and log an error if the values do not match with
what we expect.

Change-Id: I2da387b82a578e45bc974c5692e32c548acae76f
Closes-Bug: #1680367