Update the mirror metadata len correctly
While making the Sandesh interface request structure from the VIF,
mirror metadata len is not getting initizlised to zero if mirror meta
data does not exist. This is leading to previous metadata len being used
for the current metadata and display of old metadata.

As a fix, metadata length is always initialised to zero before updating
it.

Change-Id: I8ddbdf669abeec47be9f7be021f59c39cf3f0d2d
closes-bug: #1673735