Reset the reach_nh pointer to original value
If the composite nexthop does not have any component nexthops,
we set the reach_nh to discard. One side effect of this behavior
is that multicast ARPs are dropped, since ARP processing happens
in the composite nexthop handling. Hence, retain the composite
nexthop functionality, even if there are no component nexthops.

As part of this change, we also make sure that if a change in
any nexthop fails, we do not disrupt the way it works.

Change-Id: Ic9ea091e3f0cbe659389bc57bea2a1e214bbd771
Closes-Bug: #1686452
(cherry picked from commit d686b42e6c570a90d8356ffe91e1261e71661c9a)