DPDK: Fix for packet corruption when using namespace drivers
In case of LBAAS which uses DPDK af_packet driver, we were using
a framesize which is greater than the mbuf size to initialize the
SG buffer in the driver. This resulted in corruption.
Fix is to use the framsize which is a multiple of pagesize which is
lesser than the mbuf size

Change-Id: I6a678b523d866a7fdb0e89f18f855895c5391c06
Fixes-Bug: 1686273