Unset GRO flag, if GRO not invoked
Due to bug 1577473, we do not invoke GRO on a packet if Tx port
mirroring is enabled on the interface. If the packet is already marked
for GRO and if we do not subject the packet for GRO, the VP_FLAG_GRO
will not be unset. While transmitting this packet on interface, to ensure
that the tx counters are incremented only once, packet with GRO flag
disables the statistics. Due to this, tx counters are not getting
incremented when Tx port mirroring is enabled.

As a fix, GRO flag is unset unconditionally after GRO invocation checks

Change-Id: I3e5c1c2583d82e9eabfdc6dcda0eb022c9f4f422
closes-bug: #1672534