DPDK: Fix for packet corruption when using namespace drivers
In case of LBAAS which uses DPDK af_packet driver, we were using
a framesize which is greater than the mbuf size to initialize the
SG buffer in the driver.  This resulted in corruption.
Fix is to use the framsize which is a multiple of pagesize which is
lesser than the mbuf size

Change-Id: Ibf4dcacb84d398ebe56b36e935b3f78008779515
Closes-Bug: 1686273