Flow lookup for packets to/from vhost interface
For packets that come from fabric and destined to vhost's IP, if
policy needs to be applied, the lookup should not subject packets
that encapsulate overlay traffic. Overlay traffic will be governed
by its own rules and not the rules of fabric. Since we do not
differentiate the packets before subjecting it to flow lookup
(whether lookup needs to be done or not are completely determined
by the flags in the NextHop for packets destined to the NextHop or
by flags on the interface from which the packet came from. In this
case, the policy lookup is governed by the NextHop flag), with the
current implementation, such packets will also be subjected to flow
lookup.

In order to not subject encapsulated packets to policy, identify
such packets in the flow block and return FORWARD

Change-Id: I51050b69c89877becf77e901c970dd4c26cb1674
Partial-Bug: #1691023