vr_schedule_work() in dpdk
Currently hash tables defer a deletion of hash entry using a work queue,
in a different CPU.  There is no corresponding functionality in dpdk,
for this and work routine is directly invoked. This results in Hash
bucket being manipulated parallelly and resutls in crashes.

As a short cut, these Vrouter work queue routines are overloaded in dpdk
using RCU primities, which helps in the call back being invoked in a
single Cpu, which makes the hash entry manipulation from a single CPU.

This fix will be removed once we have proper fix in hash tables which
takes care of both kernel mod and dpdk.

Change-Id: Ide0442f9a253fe5eb3e716a3b7acf101a5799f4d
closes-bug: #1671324