DPDK 1604: Fix for ssh to link local not working
The checksum of packets from VM to host needs to be
calculated in software when we use tapdev. In case of
knidev, the kernel kni driver would mark the packets as
checksum unnecessary and the tcp stack would not verify
the checksum. The same is not true for tapdev. Hence we
need to compute the checksum in software.

Change-Id: I9e8787e03f85c30d015ed079a4cd849c58c8d570
Closes-Bug: 1693958
(cherry picked from commit de93ec056629d24fc21cceba7b191e0f1800ad66)