Mark the mirroring packet type correctly
Riht now the mirroring packet is marked as IP6oIP if the analyzer is V6.
If the analyzer is in same compute node as source/destination, while
transmitting this packet out, it gets treated differently. In particular
the outer IP header as such is V6 and as it is marked Ip6oIP it gets
greated as IP and results in wrong outer header handling.
To avoid this, packet in mirroring is always marked as IP or IP6.
Addition of V4 tunnel later, if exists, takes care of marking the packet
appropriately.

If analyzer is V6, the UDP ports are wrongly retrieved as V4 ports from
union. This is resulting in packet going out to different destination ports.
This also is corrected.

Conflicts:
	dp-core/vr_nexthop.c

Change-Id: I97b8f395c01786b8b98df14a56032eaf320f33b8
closes-bug: #1694414