Dont forward a packet back to same tunnel
In the following transient condition in Vrouter packet is
transmitted back to same Tunnel from which it is received.

Tor bridge entry times out, but this Evpn update is not yet received
by TSN Vrouter. As mac entry is timed out in Tor, any packet destined to
this Mac becomes unknown unicast in Tor and flooded to TSN(along with
other ports of Tor - on which the destination Mac is really lying). In TSN after
decapsulating the Vxlan packet, the bridge lookup potentially point
back to same Tor (as Vxlan tunnel NH), as Evpn delete update is not yet
received by Vrouter. This potentially can cause a loop.

As a fix, the outer tunnel source IP from which the packet is received
is validated against destination tunnel IP to which this packet is about to be
forwarded. If same, packet is dropped by incrementing a new counter.
"dropstats" has been updated to display this new counter. The fix has
been applied to Mpls tunnels as well to make it generic.

Conflicts:
	dp-core/vr_nexthop.c
	dp-core/vr_stats.c
	include/vr_packet.h
	sandesh/vr.sandesh
	utils/vr_util.c

Change-Id: Ia75097376b741d0e75cdf682075cd929003e4994
closes-bug: #1691647