Correct the return status of ENOSPC when flow table FULL
When the Agent adds a flow entry to Vrouter kernel module
(in particular reverse flow entry as forward flow entry is
 always added by data path) and if the flow table is full,
Agent expects the return status as ENOSPC. But Vrouter is
returing EEXISTS wrongly. This is because, vr_find_free_entry()
is manipulating the flow index to 0 (which is a valid flow index)
even when the free entry is not found.

As a fix, the returned flow index is not manipulated if the free
flow entry is not found

Change-Id: Ia8e68006fdea1f36fedf5b99ed9cc7884e04e3b9
closes-bug: #1699425