Choosing the correct source IP for mirroring
Currently while adding the mirroring UDP header, source IP of the packet
is chosen as mirrored packet's source IP. This works as long as the
source and destination VN's are same. Also only V4 packets can be
mirrored to V4 mirroring server and v6 pakets can be mirrored to V6
server.
To address these issues, VMI IP and IPV6 addresses are stored in VIF
directly. If the packet is mirrored because of Rx port mirroring,
ingress VIF's IP address is used as source IP of UDP tunnel. If tx port
mirroring, IP of VIF on which the packet is transmitted to VM (nh->vif) is used as
source IP of UDP tunnel. If the mirroring server is V4, vif's IPV4
address is used and if serveris V6, vif's V6 IP address is used.

Picking the ingress VIF of the vr_packet is not an issue as vp_if field
of the vr_packet always provides the ingress VIF and  is always
accessible all the time. nh->vif is not avaialble in the packet and
hence its interface index is stored in forwarding meta data (fmd). As
fmd already contains a field called fmd_mirroring_vlan, it has been
reused to store the interface index as well. fmd_mirroring_vlan is only
used incase of NIC assisted mirroring and a given packet is never marked
for both NIC assisted mirroring and regular mirroring.
fmd_mirroring_vlan has been renamed to fmd_mirroring_data and has been
encoded with MSB bits to identify whether stored value is Vlan or
Interface index. Along with interface index, whether it is ingress VIF
or egress VIF is also marked in the same field.

Change-Id: Icc25841ad2deb487306e3a810d4de242177f435e
closes-bug: #1687720