Dont update flow index if find flow fails
While searching the flow table, if the flow lookup fails,
the flow index is still modified though search has failed.
This is resulting in caller returning  a different error status
though the lookup resulted in failure. As a fix, the flow index
is updated only if the flow lookup succeds

This fix is continuation of the fix https://review.opencontrail.org/#/c/33404/

closes-bug: #1699425

Change-Id: I5acd2b4da30304f889ed6ba790211b1534f54cfa