Add VM GREoIPv4 packets to traffic distribution mechanism
When GRE over IPv4 packets are coming from VM they are not
being addressed for rss hash recalculation in functions
    dpdk_mbuf_parse_and_hash_packets()
    dpdk_mbuf_rss_hash()

that leads to distribute such packets on the same core for
routing if such traffic pps is higher than single core performance
then it will lead to traffic drops and this core became a botleneck.

Additionaly in dpdk_mbuf_rss_hash() if gre key is present
in the packet this gre key is added to rss hash.
That allows to distribute traffic segregated by gre key not
only Src IPv4 Dst IPv4.

Change-Id: I1542e5b68e1610a90614d33a140080307772341f
Closes-Bug: #1706344