Underlay ARP handling
Following are the changes as part of this underlay ARP handling (Agent
behaviour is also listed for the complete understanding of the
packet processing)

Agents Changes:

1) VMI of VM's in IP Frabric VN are pushed to unrestricted proxy ARP by Agent

2) /32 routes of VM in IP Fabric VN that are hosted on that particular
compute node are added to Inet route table with "P" bit and without stitching

3) /32 routes of VM in IP Fabric VN that are hosted on different compute
nodes are added to Inet route table with "P" bit and as well with
stitching. Stitched MAC address would be that of other compute node or
default gw (if the two compute nodes are present in different subnets)

4) Compute node /32 IP are added in Inet table without any bits

Vrouter changes:

A) ARP request packets on Vhost and Fabric are no more cross connected
and are subjected to regular ARP processing

B) There would be a IP lookup in the inet table of default VRF for
destintio IP of ARP packet. The type of NH and the route flags are used
for the below processing

    1) ARP requests from Host OS:
        a) if destination IP is a VM in hosted in same compute node, due to
           (2), arp response would be sent with Vhost MAC address

        b) if destination IP is a VM hosted in different compute node,
           due to (3), ARP response would be sent with stitched mac of
           the Inet route

        c) If destinatio IP is another compute node, due to (4), packet
           would be cross connected to Fabric interface

    2)  ARP requests from VM:
        Due to (1) all ARP requests from VM are answered with Vrouter's
        MAC

    3)  ARP requests from Fabric:
         a) if destination is Vhost IP(or other any other compute node
             IP), due to (4), request would be cross connected to Vhost
             interface, leaving the request to be answered or rejected
             by HostOS

         b) if destination is for VM hosted on this compute node, due to
            (2), ARP response would be sent with Vhost Mac address

         c) if destination is for VM not hosted on this compute node,
            due to (3), ARP request would be dropped, leaving the request
            to be answered by the respective compute node

C) ARP response:
       a) All ARP response packets from Vhost are cross connected
       b) ARP response packets from Fabric are cross connected to
          vhost and trapped to agent as well

D) All ARP responses on unrestricted Proxy ARP interface are transmitted
   back on the interface without a bridge lookup

E) ARP requests from unrestricted proxy ARP interface are subjected to
   ARP processing without L2 multicast NH

F) In Encap nexthop of Vhost interface, for IP/IP6 packets from vhost
   interface, L2 header from the HostOS is retained without adding Agent
   constructed rewrite data. If the packet is tunneled one - due to
   Vhost speaking to overlay - the rewrite data that is provided by
   Agent is applied

Change-Id: I33cd97f648ba21923164a6cd57627d63d19d7486
partial-bug: #1702837